% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outbreakHNB.R
\name{outbreakHNB}
\alias{outbreakHNB}
\title{Fit an outbreak detection Hurdle negative binomial outbreak model}
\usage{
outbreakHNB(
  cases,
  pop = NULL,
  covariates_count = NULL,
  covariates_zero = NULL,
  beta_init = NULL,
  delta_init = NULL,
  r_init = NULL,
  beta_prior_mean = 0,
  beta_prior_sd = 10,
  delta_prior_mean = 0,
  delta_prior_sd = 10,
  r_prior_shape = 1,
  r_prior_rate = 1,
  p_priors = 1,
  n_iter = 1e+05,
  n_burnin = 10000,
  n_chains = 3,
  n_thin = 1,
  save_params = c("beta", "delta", "r", "Z"),
  dates = NULL,
  plot_Z = FALSE
)
}
\arguments{
\item{cases}{Integer or numeric vector of observed case counts (length N).}

\item{pop}{(Optional) Numeric vector of population offsets (length N). If NULL, offset = 1.}

\item{covariates_count}{(Optional) Data.frame or matrix of covariates for the count model (N x p_c).}

\item{covariates_zero}{(Optional) Data.frame or matrix of covariates for the Hurdle model (N x p_z).}

\item{beta_init}{(Optional) List of length \code{n_chains} giving initial values for β (each a vector of length p_c+1).}

\item{delta_init}{(Optional) List of length \code{n_chains} giving initial values for δ (each a vector of length p_z+1).}

\item{r_init}{(Optional) Numeric vector of length \code{n_chains} giving initial values for the NB dispersion parameter.}

\item{beta_prior_mean}{Prior mean for beta coefficients of the Negative binomial part (default = 0).}

\item{beta_prior_sd}{Prior SD   for beta coefficients of the Negative binomial part (default = 10).}

\item{delta_prior_mean}{Prior mean for delta coefficients of the Hurdle part (default = 0).}

\item{delta_prior_sd}{Prior SD   for delta coefficients of the Hurdle part (default = 10).}

\item{r_prior_shape}{Shape parameter of a prior on r (default = 1).}

\item{r_prior_rate}{Rate  parameter of b prior on r (default = 1).}

\item{p_priors}{Alpha parameters for the binomial priors on p00 and p11 (default = 1).}

\item{n_iter}{Total number of MCMC iterations per chain (default = 100000).}

\item{n_burnin}{Number of burn-in iterations (default = 10000).}

\item{n_chains}{Number of MCMC chains (default = 3).}

\item{n_thin}{Thinning interval for MCMC samples (default = 1).}

\item{save_params}{Character vector of parameter names to save (must include "Z").}

\item{dates}{(Optional) Vector of Date or POSIX dates for plotting Z; if NULL, uses index 1:N.}

\item{plot_Z}{Logical; if TRUE, returns a ggplot2 object of the posterior mean Z over time.}
}
\value{
A list with MCMC summary, samples, DIC, WAIC, and plot of the probability of being in an epidemic state.
}
\description{
Fit an outbreak detection Hurdle negative binomial outbreak model
}
