% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsesurv-package.R
\docType{package}
\name{sparsesurv-package}
\alias{sparsesurv}
\alias{sparsesurv-package}
\title{sparsesurv: Forecasting and Early Outbreak Detection for Sparse Count Data}
\description{
Tools for fitting, forecasting, and early detection of outbreaks in sparse surveillance count time series. Supports negative binomial (NB), self-exciting NB, GARMA-NB, zero-inflated NB (ZINB), self-exciting ZINB, GARMA-ZINB, and hurdle formulations. Climatic and environmental covariates can be included in the regression component and/or the zero-modified components. Includes outbreak-detection algorithms for NB, ZINB, and hurdle models, with utilities for prediction and diagnostics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YOUR_GH_USER/sparsesurv}
  \item \doi{10.5281/zenodo.xxxxxxx}
  \item Report bugs at \url{https://github.com/YOUR_GH_USER/sparsesurv/issues}
}

}
\author{
\strong{Maintainer}: Alexandros Angelakis \email{alexandros.angelakis@swisstph.ch}

}
\keyword{internal}
